+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Support for GDK101 gamma radiation sensor by Petr Novacek (#18390)
 - Matter support in now stabilized for Apple and Google (not tested with Alexa)
 - Berry `instrospect.name()` to get names of functions, modules and classes (#18422)
+- Berry add `searchall()` and `matchall()` to `re` module and pre-compiled patterns
 
 ### Breaking Changed